chore: divide checkout tests for better debugging
